+++ src/index.ts
@@ -1,9 +1,9 @@
 export * from './types';
-export * as body from './plugins/body';
-export * as server from './server';
-export * as client from './client';
-export * as router from './router';
+export * from './server';
+export * from './client';
+export * from './constants';
+export * from './controllers';
+export * from './routes/files';
+export * from './plugins/request';
+export * from './plugins/response';
 export * as cookies from './plugins/cookies';
-export * as respond from './plugins/respond';
-export * as constants from './constants';
-export { files } from './handlers/files';