+++ src/plugins/body/json.ts
@@ -1,4 +1,3 @@
-import { ContentTypeValue } from '../../constants';
 import {
   constants,
   Http2Stream,
@@ -6,24 +5,26 @@ import {
   IncomingHttpStatusHeader,
 } from 'http2';
 
+import { ContentType } from '../../constants';
+
 /** Parses request body as `JSON` format. */
-export function json<T extends object>(
+export const json = async <T extends object>(
   stream: Http2Stream,
   headers: IncomingHttpHeaders & IncomingHttpStatusHeader
-): Promise<T> {
-  const contentTypeHeaderValue =
-    headers[constants.HTTP2_HEADER_CONTENT_TYPE] ?? '';
+): Promise<T> =>
+  new Promise<string>((resolve, reject) => {
+    const contentTypeHeaderValue =
+      headers[constants.HTTP2_HEADER_CONTENT_TYPE] ?? '';
 
-  return new Promise<string>((resolve, reject) => {
     const chunks: Array<string> = [];
 
     if (
       // Value of Content-Type header can contain charset field.
-      !contentTypeHeaderValue.includes(ContentTypeValue.JSON)
+      !contentTypeHeaderValue.includes(ContentType.JSON)
     ) {
       reject(
         new Error(
-          `Unexpected Content-Type - "${contentTypeHeaderValue}". Expected "${ContentTypeValue.JSON}"`
+          `Unexpected Content-Type - "${contentTypeHeaderValue}". Expected "${ContentType.JSON}"`
         )
       );
     }
@@ -34,4 +35,3 @@ export function json<T extends object>(
       .on('data', chunks.push.bind(chunks))
       .on('end', () => resolve(chunks.join('')));
   }).then<T>(JSON.parse);
-}