+++ src/plugins/body/text.ts
@@ -1,8 +1,8 @@
-import type { Http2Stream } from 'http2';
+import { Http2Stream } from 'http2';
 
-/** Parses reauest body as `UTF8` string. */
-export async function text(stream: Http2Stream): Promise<string> {
-  return new Promise<string>((resolve, reject) => {
+/** Parses request body as `UTF8` string. */
+export const text = async (stream: Http2Stream): Promise<string> =>
+  new Promise<string>((resolve, reject) => {
     const chunks: Array<string> = [];
 
     stream
@@ -11,4 +11,3 @@ export async function text(stream: Http2Stream): Promise<string> {
       .on('data', chunks.push.bind(chunks))
       .on('end', () => resolve(chunks.join('')));
   });
-}