+++ src/plugins/body/urlencoded.ts
@@ -1,5 +1,3 @@
-import { parse } from 'querystring';
-import { ContentTypeValue } from '../../constants';
 import {
   constants,
   Http2Stream,
@@ -7,27 +5,29 @@ import {
   IncomingHttpStatusHeader,
 } from 'http2';
 
+import { ContentType } from '../../constants';
+
 /**
  * Parses request with `application/x-www-form-urlencoded`.
  * @returns object with key/value pairs.
  */
-export async function urlencoded<T extends Record<string, string>>(
+export const urlencoded = async <T extends Record<string, string>>(
   stream: Http2Stream,
   headers: IncomingHttpHeaders & IncomingHttpStatusHeader
-): Promise<T> {
-  const contentTypeHeaderValue =
-    headers[constants.HTTP2_HEADER_CONTENT_TYPE] ?? '';
+): Promise<T> =>
+  new Promise<string>((resolve, reject) => {
+    const contentTypeHeaderValue =
+      headers[constants.HTTP2_HEADER_CONTENT_TYPE] ?? '';
 
-  return new Promise<string>((resolve, reject) => {
     const chunks: Array<string> = [];
 
     if (
       // Content-Type can contain charset field.
-      !contentTypeHeaderValue.includes(ContentTypeValue.FORM_URLENCODED)
+      !contentTypeHeaderValue.includes(ContentType.FORM_URLENCODED)
     ) {
       reject(
         new Error(
-          `Unexpected Content-Type - "${contentTypeHeaderValue}". Expected "${ContentTypeValue.FORM_URLENCODED}"`
+          `Unexpected Content-Type - "${contentTypeHeaderValue}". Expected "${ContentType.FORM_URLENCODED}"`
         )
       );
     }
@@ -37,5 +37,9 @@ export async function urlencoded<T extends Record<string, string>>(
       .on('error', reject)
       .on('data', chunks.push.bind(chunks))
       .on('end', () => resolve(chunks.join('')));
-  }).then((result) => ({ ...parse(result) } as T));
-}
+  }).then((result) =>
+    Array.from(new URLSearchParams(result).entries()).reduce(
+      (params, [name, value]) => ({ ...params, [name]: value }),
+      {} as T
+    )
+  );